AASIST3/requirements.txt
accelerate==1.7.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
comet_ml==3.49.12
configobj==5.0.9
dulwich==0.23.2
everett==3.1.0
filelock==3.18.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
hf-xet==1.1.3
huggingface-hub==0.32.4
hydra-core==1.3.2
idna==3.10
importlib-metadata==6.11.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
julius==0.2.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.48.1
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
packaging==23.2
pandas==2.2.3
Pillow==9.5.0
platformdirs==4.3.8
primePy==1.3
protobuf==4.25.8
psutil==7.0.0
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
Pympler==1.1
python-box==6.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.26.0
safetensors==0.5.3
semantic-version==2.10.0
sentry-sdk==2.29.1
setproctitle==1.3.6
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
soundfile==0.12.1
streamlit==1.25.0
sympy==1.13.1
tenacity==8.5.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.6.0
torcheval==0.0.7
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.4
triton==3.2.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==4.3.1
urllib3==2.4.0
validators==0.35.0
wandb==0.20.1
watchdog==6.0.0
wrapt==1.17.2
wurlitzer==3.1.1
zipp==3.23.0
